แก้ไขหน้าแรกให้เป็นหน้า Login แก้ไขหน้า Google-login เพิ่ม link register ในหน้า Login_Activity
