+++ app/src/main/res/layout/activity_main.xml
@@ -38,12 +38,7 @@
             android:layout_gravity="center"
             android:text="Register_page" />
 
-        <ImageView
-            android:id="@+id/img_google"
-            android:layout_width="250dp"
-            android:layout_height="150dp"
-            android:layout_gravity="center"
-            app:srcCompat="@drawable/sso_google" />
+
     </LinearLayout>
 
 </LinearLayout>
